# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Flow.",
        "title": "Flow API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/flows": {
            "get": {
                "description": "Get public pin by ID or private pin if user its author.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pin CRUD"
                ],
                "summary": "Get pin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "requested pin",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid query parameter [id]",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access to private pin is forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "no pin with given id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "untracked error: ${error}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update certain pin's fields. Authorization is required. Pin must belong to user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pin CRUD"
                ],
                "summary": "Update pin",
                "parameters": [
                    {
                        "description": "pin ID",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "text header",
                        "name": "header",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "text description",
                        "name": "description",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "privacy setting",
                        "name": "is_private",
                        "in": "body",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "no fields to update",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "user is not authorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access to private pin is forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "no pin with given id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "untracked error: ${error}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create pin with given header, description and private setting. Authorization is required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pin CRUD"
                ],
                "summary": "Create pin",
                "parameters": [
                    {
                        "type": "file",
                        "description": "pin image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "text header",
                        "name": "header",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "text description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "privacy setting",
                        "name": "is_private",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid image extension",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "user is not authorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "untracked error: ${error}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete pin by ID. Authorization is required. Pin must belong to user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pin CRUD"
                ],
                "summary": "Delete pin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pin to delete",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid query parameter [id]",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "access to private pin is forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "no pin with given id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "untracked error: ${error}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}